backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
anthropic==0.28.0
vosk==0.3.44
python-docx==1.1.2
pydantic==2.7.1
aiofiles==23.2.1
httpx==0.27.0
pydub==0.25.1
